consumer/requirements.txt
google-cloud-pubsub==2.27.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.1
google-cloud-secret-manager==2.18.0
fastavro==1.9.4
pandas==2.2.0
pyarrow==15.0.0
